Common: Rename i_event_serialize{,r}.py
