+++ monkey/common/event_serializers/__init__.py
@@ -1,3 +1,3 @@
-from .i_event_serialize import IEventSerializer, JSONSerializable
+from .i_event_serializer import IEventSerializer, JSONSerializable
 from .event_serializer_registry import EventSerializerRegistry
 from .pydantic_event_serializer import PydanticEventSerializer
