+++ monkey/common/event_serializers/i_event_serializer.py
